refactor : EmailAuthService 삭제 및 해당 로직 ConfirmationTokenService로 이동
